requirements.txt
streamlit
groq
python-dotenv
pypdf
reportlab